requirements.txt
aiogram==3.3.0
python-telegram-bot==20.7
pyTelegramBotAPI==4.14.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.2.1
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
Werkzeug==3.0.0
